fixed utils and stake hist
