+++ program/src/helpers/mod.rs
@@ -1,7 +1,5 @@
-pub mod split;
-pub mod sysvar;
+pub mod sysvar_interface_stake_history;
 pub mod sysvar_stake_history;
 
-pub use split::*;
-pub use sysvar::*;
+pub use sysvar_interface_stake_history::*;
 pub use sysvar_stake_history::*;
